Import amule 2.2.0~20080309 tarball.

+++ src/BarShader.cpp
@@ -1,7 +1,7 @@
 //
 // This file is part of the aMule Project.
 //
-// Copyright (c) 2003-2006 aMule Team ( admin@amule.org / http://www.amule.org )
+// Copyright (c) 2003-2008 aMule Team ( admin@amule.org / http://www.amule.org )
 // Copyright (c) 2002 Merkur ( devs@emule-project.net / http://www.emule-project.net )
 //
 // Any parts of this program derived from the xMule, lMule or eMule project,
@@ -23,9 +23,6 @@
 // Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301, USA
 //
 
-#include <cmath>
-#include <algorithm>		// Needed for std::min
-#include <wx/gdicmn.h>
 #include <wx/dc.h>
 #include "Color.h"		// Needed for RGB
 
@@ -63,7 +60,7 @@ void CBarShader::Reset()
 }
 
 
-void CBarShader::SetFileSize(uint32 fileSize)
+void CBarShader::SetFileSize(uint64 fileSize)
 {
 	m_FileSize = fileSize;
 	Reset();
@@ -122,7 +119,7 @@ void CBarShader::BuildModifiers()
 }
 
 
-void CBarShader::FillRange(uint32 start, uint32 end, const uint32 color)
+void CBarShader::FillRange(uint64 start, uint64 end, const uint32 color)
 {
 	// Sanity check
 	wxASSERT( start <= end );
@@ -172,7 +169,7 @@ void CBarShader::Draw( wxDC* dc, int iLeft, int iTop, bool bFlat )
 	// This modifier is multipled with sizes to allow for better handling of small ranges
 	const uint64 MOD = 1000;
 	// This is the number of bits each pixel should contain.
-	const uint64 bitsPerPixel = ((uint64)m_FileSize * MOD) / (uint64)m_Width;
+	const uint64 bitsPerPixel = (m_FileSize * MOD) / (uint64)m_Width;
 	
 	// The initial values for partial pixel drawing
 	uint64 curPixel = 0;
@@ -283,4 +280,4 @@ void CBarShader::FillRect(wxDC *dc, const wxRect& rectSpan, uint32 color, bool b
 
 	dc->SetBrush(wxNullBrush);
 }
-
+// File_checked_for_headers
